+++ sdk-java/src/main/java/ly/count/sdk/java/internal/CrashImplCore.java
@@ -19,18 +19,19 @@ import ly.count.sdk.java.Crash;
  */
 
 public class CrashImplCore implements Crash, Storable {
-    private static final Log.Module L = Log.module("CrashImpl");
+    private Log L = null;
 
     private final Long id;
     private final JSONObject data;
     private Throwable throwable;
     private Map<Thread, StackTraceElement[]> traces;
 
-    protected CrashImplCore() {
-        this(DeviceCore.dev.uniformTimestamp());
+    protected CrashImplCore(Log logger) {
+        this(DeviceCore.dev.uniformTimestamp(), logger);
     }
 
-    protected CrashImplCore(Long id) {
+    protected CrashImplCore(Long id, Log logger) {
+        this.L = logger;
         this.id = id;
         this.data = new JSONObject();
         this.add("_nonfatal", true);
@@ -54,7 +55,7 @@ public class CrashImplCore implements Crash, Storable {
     @Override
     public CrashImplCore addTraces(Thread main, Map<Thread, StackTraceElement[]> traces) {
         if (traces == null) {
-            L.wtf("traces cannot be null");
+          L.e("[CrashImpl traces cannot be null");
             return this;
         } else {
             this.traces = traces;
@@ -164,7 +165,7 @@ public class CrashImplCore implements Crash, Storable {
     public List<String> getLogs() {
         try {
             String logs = this.data.getString("_logs");
-            return Utils.isEmpty(logs) ? null : Arrays.asList(logs.split("\n"));
+            return Utils.isEmptyOrNull(logs) ? null : Arrays.asList(logs.split("\n"));
         } catch (JSONException e) {
             return null;
         }
@@ -175,7 +176,7 @@ public class CrashImplCore implements Crash, Storable {
             try {
                 this.data.put(key, value);
             } catch (JSONException e) {
-                L.wtf("Couldn't add " + key + " to a crash", e);
+              L.e("[CrashImpl Couldn't add " + key + " to a crash" + e);
             }
         }
         return this;
@@ -186,7 +187,7 @@ public class CrashImplCore implements Crash, Storable {
         try {
             return data.toString().getBytes(Utils.UTF8);
         } catch (UnsupportedEncodingException e) {
-            L.wtf("UTF is not supported", e);
+          L.e("[CrashImpl UTF is not supported" + e);
             return null;
         }
     }
@@ -202,11 +203,11 @@ public class CrashImplCore implements Crash, Storable {
                     this.data.put(k, obj.get(k));
                 }
             } catch (JSONException e) {
-                L.e("Couldn't decode crash data successfully", e);
+               L.e("[CrashImpl Couldn't decode crash data successfully" + e);
             }
             return true;
         } catch (UnsupportedEncodingException e) {
-            L.wtf("Cannot deserialize crash", e);
+          L.e("[CrashImpl Cannot deserialize crash" + e);
         }
 
         return false;
@@ -229,7 +230,7 @@ public class CrashImplCore implements Crash, Storable {
     public CrashImplCore putMetricsCore(CtxCore ctx, Long runningTime) {
         String version = ctx.getConfig().getApplicationVersion();
         return add("_os", DeviceCore.dev.getOS())
-                .add("_app_version", Utils.isEmpty(version) ? "0.0" : version)
+                .add("_app_version", Utils.isEmptyOrNull(version) ? "0.0" : version)
                 .add("_os_version", DeviceCore.dev.getOSVersion())
                 .add("_ram_current", DeviceCore.dev.getRAMAvailable())
                 .add("_ram_total", DeviceCore.dev.getRAMTotal())