+++ sdk-java/src/main/java/ly/count/sdk/java/internal/Log.java
@@ -7,79 +7,29 @@ import ly.count.sdk.java.Config;
  * in time when {@link Config} is created and {@link Module}s are up.
  */
 
-public class Log extends ModuleBase {
-    private static Log instance;
+public class Log {
 
     private String tag;
-    private boolean testMode;
+    LogCallback logListener = null;
     private Config.LoggingLevel level;
 
-    public static final class Module {
-        String name;
 
-        Module(String name) {
-            this.name = name;
-        }
-
-        public void d(String message) { if(instance != null) instance.d("[" + name + "] " + message); }
-        public void d(String message, Throwable throwable) { if(instance != null) instance.d("[" + name + "] " + message, throwable); }
-        public void i(String message) { if(instance != null) instance.i("[" + name + "] " + message); }
-        public void i(String message, Throwable throwable) { if(instance != null) instance.i("[" + name + "] " + message, throwable); }
-        public void w(String message) { if(instance != null) instance.w("[" + name + "] " + message); }
-        public void w(String message, Throwable throwable) { if(instance != null) instance.w("[" + name + "] " + message, throwable); }
-        public void e(String message) { if(instance != null) instance.e("[" + name + "] " + message); }
-        public void e(String message, Throwable throwable) { if(instance != null) instance.e("[" + name + "] " + message, throwable); }
-        public void wtf(String message) { if(instance != null) instance.wtf("[" + name + "] " + message); }
-        public void wtf(String message, Throwable throwable) { if(instance != null) instance.wtf("[" + name + "] " + message, throwable); }
-    }
-
-    @Override
-    public void init(InternalConfig config) {
-        instance = this;
-
+    public Log(InternalConfig config) {
         // let it be specific int and not index for visibility
         tag = config.getLoggingTag();
         level = config.getLoggingLevel();
-        testMode = config.isTestModeEnabled();
-
-    }
-
-    @Override
-    public Integer getFeature() {
-        return CoreFeature.Logs.getIndex();
-    }
-
-    public static void deinit() {
-        instance = null;
-    }
-
-    public static Module module(String name) {
-        return new Module(name);
-    }
-
-    /**
-     * {@link Config.LoggingLevel#DEBUG} level logging
-     *
-     * @param string string to log
-     */
-    public  void d(String string) {
-        d(string, null);
+        logListener = config.getLogListener();
     }
 
     /**
      * {@link Config.LoggingLevel} level logging
      *
      * @param string string to log
-     * @param t exception to log along with {@code string}
      */
-    public void d(String string, Throwable t) {
-        if (level != null && level.prints(Config.LoggingLevel.DEBUG)) {
-            if (t == null) {
-                System.out.println("[DEBUG]\t" + tag + "\t" + string);
-            } else {
-                System.out.println("[DEBUG]\t" + tag + "\t" + string + " / " + t);
-            }
-        }
+    public void d(String string) {
+        String msg = "[DEBUG]\t" + tag + "\t" + string;
+        print(msg, Config.LoggingLevel.DEBUG);
+        informListener(msg, Config.LoggingLevel.DEBUG);
     }
 
     /**
@@ -88,23 +38,9 @@ public class Log extends ModuleBase {
      * @param string string to log
      */
     public void i(String string) {
-        i(string, null);
-    }
-
-    /**
-     * {@link Config.LoggingLevel#INFO} level logging
-     *
-     * @param string string to log
-     * @param t exception to log along with {@code string}
-     */
-    public void i(String string, Throwable t) {
-        if (level.prints(Config.LoggingLevel.INFO)) {
-            if (t == null) {
-                System.out.println("[INFO]\t" + tag + "\t" + string);
-            } else {
-                System.out.println("[INFO]\t" + tag + "\t" + string + " / " + t);
-            }
-        }
+        String msg = "[INFO]\t" + tag + "\t" + string;
+        print(msg, Config.LoggingLevel.INFO);
+        informListener(msg, Config.LoggingLevel.INFO);
     }
 
     /**
@@ -113,23 +49,9 @@ public class Log extends ModuleBase {
      * @param string string to log
      */
     public void w(String string) {
-        w(string, null);
-    }
-
-    /**
-     * {@link Config.LoggingLevel#WARN} level logging
-     *
-     * @param string string to log
-     * @param t exception to log along with {@code string}
-     */
-    public void w(String string, Throwable t) {
-        if (level != null && level.prints(Config.LoggingLevel.WARN)) {
-            if (t == null) {
-                System.out.println("[WARN]\t" + tag + "\t" + string);
-            } else {
-                System.out.println("[WARN]\t" + tag + "\t" + string + " / " + t);
-            }
-        }
+        String msg = "[WARN]\t" + tag + "\t" + string;
+        print(msg, Config.LoggingLevel.WARN);
+        informListener(msg, Config.LoggingLevel.WARN);
     }
 
     /**
@@ -138,51 +60,31 @@ public class Log extends ModuleBase {
      * @param string string to log
      */
     public void e(String string) {
-        e(string, null);
+        String msg = "[ERROR]\t" + tag + "\t" + string;
+        print(msg, Config.LoggingLevel.ERROR);
+        informListener(msg, Config.LoggingLevel.ERROR);
     }
 
     /**
-     * {@link Config.LoggingLevel#ERROR} level logging
+     * {@link Config.LoggingLevel#VERBOSE} level logging
      *
      * @param string string to log
-     * @param t exception to log along with {@code string}
      */
-    public void e(String string, Throwable t) {
-        if (level != null && level.prints(Config.LoggingLevel.ERROR)) {
-            if (t == null) {
-                System.out.println("[ERROR]\t" + tag + "\t" + string);
-            } else {
-                System.out.println("[ERROR]\t" + tag + "\t" + string + " / " + t);
-            }
+    public void v(String string) {
+        String msg = "[VERBOSE]\t" + tag + "\t" + string;
+        print(msg, Config.LoggingLevel.VERBOSE);
+        informListener(msg, Config.LoggingLevel.VERBOSE);
+    }
+
+    private void print(String msg, Config.LoggingLevel level) {
+        if (level != null && level.prints(level)) {
+            System.out.println(msg);
         }
     }
 
-    /**
-     * {@link Config.LoggingLevel#ERROR} (Android wtf) level logging which throws an
-     * exception when {@link Config#testMode} is enabled.
-     *
-     * @param string string to log
-     * @throws IllegalStateException when {@link Config#testMode} is on
-     */
-
-    public void wtf(String string) {
-        wtf(string, null);
-    }
-
-    /**
-     * {@link Config.LoggingLevel#ERROR} (Android wtf) level logging which throws an
-     * exception when {@link Config#testMode} is enabled.
-     *
-     * @param string string to log
-     * @param t exception to log along with {@code string}
-     */
-    public void wtf(String string, Throwable t) {
-        if (level != null && level != Config.LoggingLevel.OFF) {
-            if (t == null) {
-                System.out.println("[WTF]\t" + tag + "\t" + string);
-            } else {
-                System.out.println("[WTF]\t" + tag + "\t" + string + " / " + t);
-            }
+    private void informListener(String msg, Config.LoggingLevel level) {
+        if (logListener != null) {
+            logListener.LogHappened(msg, level);
         }
     }
 }