+++ sdk-java/src/main/java/ly/count/sdk/java/internal/UserEditorImpl.java
@@ -17,7 +17,7 @@ import ly.count.sdk.java.User;
 import ly.count.sdk.java.UserEditor;
 
 public class UserEditorImpl implements UserEditor {
-    private static final Log.Module L = Log.module("UserEditorImpl");
+    private  Log L = null;
 
     static class Op {
         static final String INC = "$inc";
@@ -136,7 +136,7 @@ public class UserEditorImpl implements UserEditor {
     private final List<Op> ops;
     private final List<String> cohortsToAdd, cohortsToRemove;
 
-    UserEditorImpl(UserImpl user) {
+    UserEditorImpl(UserImpl user, Log logger) {
         this.user = user;
         this.sets = new HashMap<>();
         this.ops = new ArrayList<>();
@@ -202,7 +202,7 @@ public class UserEditorImpl implements UserEditor {
                         user.picture = (byte[]) value;
                         changes.put(PICTURE_PATH, PICTURE_IN_USER_PROFILE);
                     } else {
-                        L.wtf("Won't set user picture (must be of type byte[])");
+                       L.e("[UserEditorImpl Won't set user picture (must be of type byte[])");
                     }
                     break;
                 case PICTURE_PATH:
@@ -215,10 +215,10 @@ public class UserEditorImpl implements UserEditor {
                             user.picturePath = new URI((String) value).toString();
                             changes.put(PICTURE_PATH, user.picturePath);
                         } catch (URISyntaxException e) {
-                            L.wtf("Supplied picturePath is not parsable to java.net.URI");
+                           L.e("[UserEditorImpl Supplied picturePath is not parsable to java.net.URI");
                         }
                     } else {
-                        L.wtf("Won't set user picturePath (must be String or null)");
+                       L.e("[UserEditorImpl Won't set user picturePath (must be String or null)");
                     }
                     break;
                 case GENDER:
@@ -228,13 +228,13 @@ public class UserEditorImpl implements UserEditor {
                     } else if (value instanceof String) {
                         User.Gender gender = User.Gender.fromString((String) value);
                         if (gender == null) {
-                            L.wtf("Cannot parse gender string: " + value + " (must be one of 'F' & 'M')");
+                           L.e("[UserEditorImpl Cannot parse gender string: " + value + " (must be one of 'F' & 'M')");
                         } else {
                             user.gender = gender;
                             changes.put(GENDER, user.gender.toString());
                         }
                     } else {
-                        L.wtf("Won't set user gender (must be of type User.Gender or one of following Strings: 'F', 'M')");
+                       L.e("[UserEditorImpl Won't set user gender (must be of type User.Gender or one of following Strings: 'F', 'M')");
                     }
                     break;
                 case BIRTHYEAR:
@@ -246,10 +246,10 @@ public class UserEditorImpl implements UserEditor {
                             user.birthyear = Integer.parseInt((String) value);
                             changes.put(BIRTHYEAR, user.birthyear);
                         } catch (NumberFormatException e) {
-                            L.wtf("user.birthyear must be either Integer or String which can be parsed to Integer", e);
+                           L.e("[UserEditorImpl user.birthyear must be either Integer or String which can be parsed to Integer" + e);
                         }
                     } else {
-                        L.wtf("Won't set user birthyear (must be of type Integer or String which can be parsed to Integer)");
+                       L.e("[UserEditorImpl Won't set user birthyear (must be of type Integer or String which can be parsed to Integer)");
                     }
                     break;
                 case LOCALE:
@@ -289,7 +289,7 @@ public class UserEditorImpl implements UserEditor {
                             user.custom.put(key, value);
                         }
                     } else {
-                        L.wtf("Type of value " + value + " '" + value.getClass().getSimpleName() + "' is not supported yet, thus user property is not stored");
+                       L.e("[UserEditorImpl Type of value " + value + " '" + value.getClass().getSimpleName() + "' is not supported yet, thus user property is not stored");
                     }
                     break;
             }
@@ -435,11 +435,11 @@ public class UserEditorImpl implements UserEditor {
                 try {
                     return set(LOCATION, Double.valueOf(comps[0]) + "," + Double.valueOf(comps[1]));
                 } catch (Throwable t){
-                    L.wtf("Invalid location format: " + location, t);
+                   L.e("[UserEditorImpl Invalid location format: " + location + " " + t);
                     return this;
                 }
             } else {
-                L.wtf("Invalid location format: " + location);
+               L.e("[UserEditorImpl Invalid location format: " + location);
                 return this;
             }
         } else {
@@ -491,7 +491,7 @@ public class UserEditorImpl implements UserEditor {
     public UserEditor setOnce(String key, Object value) {
         L.d("setOnce: key " + key + " value " + value);
         if (value == null) {
-            L.wtf("$setOnce operation operand cannot be null: key " + key);
+           L.e("[UserEditorImpl $setOnce operation operand cannot be null: key " + key);
             return this;
         } else {
             return setCustomOp(Op.SET_ONCE, key, trimValue(key, value));
@@ -502,7 +502,7 @@ public class UserEditorImpl implements UserEditor {
     public UserEditor pull(String key, Object value) {
         L.d("pull: key " + key + " value " + value);
         if (value == null) {
-            L.wtf("$pull operation operand cannot be null: key " + key);
+           L.e("[UserEditorImpl $pull operation operand cannot be null: key " + key);
             return this;
         } else {
             return setCustomOp(Op.PULL, key, value);
@@ -513,7 +513,7 @@ public class UserEditorImpl implements UserEditor {
     public UserEditor push(String key, Object value) {
         L.d("push: key " + key + " value " + value);
         if (value == null) {
-            L.wtf("$push operation operand cannot be null: key " + key);
+           L.e("[UserEditorImpl $push operation operand cannot be null: key " + key);
             return this;
         } else {
             return setCustomOp(Op.PUSH, key, value);
@@ -524,7 +524,7 @@ public class UserEditorImpl implements UserEditor {
     public UserEditor pushUnique(String key, Object value) {
         L.d("pushUnique: key " + key + " value " + value);
         if (value == null) {
-            L.wtf("pushUnique / $addToSet operation operand cannot be null: key " + key);
+           L.e("[UserEditorImpl pushUnique / $addToSet operation operand cannot be null: key " + key);
             return this;
         } else {
             return setCustomOp(Op.PUSH_UNIQUE, key, value);
@@ -555,7 +555,7 @@ public class UserEditorImpl implements UserEditor {
     public User commit() {
         L.d("commit");
         if (SDKCore.instance == null) {
-            L.wtf("Countly is not initialized");
+           L.e("[UserEditorImpl Countly is not initialized");
             return null;
         }
 
@@ -581,7 +581,7 @@ public class UserEditorImpl implements UserEditor {
                         try {
                             params.add(PICTURE_PATH, changes.getString(PICTURE_PATH));
                         } catch (JSONException e) {
-                            L.w("Won't send picturePath", e);
+                            L.w("Won't send picturePath" + e);
                         }
                     }
                     if (cohortsAdded.size() > 0) {
@@ -608,7 +608,7 @@ public class UserEditorImpl implements UserEditor {
             SDKCore.instance.onUserChanged(user.ctx, changes, cohortsAdded, cohortsRemoved);
 
         } catch (JSONException e) {
-            L.wtf("Exception while committing changes to User profile", e);
+           L.e("[UserEditorImpl Exception while committing changes to User profile" + e);
         }
 
         sets.clear();